lru: Implement Map type
Map is like Cache but records both a key and a value.

A new Get method is added to receive the value for a key, if present.  When
found, the item is moved to the front of the most-recently-used list.  Another
method Hit is similar, marking the item (if any) as the most recently used,
but discards the value.