Fix acquire ordering on DB, addresses mutexes
This resolves a deadlock caused by interlaced ordering acquiring the internal
bbolt mutexes and the wallet's addresses mutex (sometimes acquiring bbolt
mutexes with the address mutex locked, other times not).  This is done by
rewriting nextAddress to no longer take a callback function to persist the
address to the database (at times, opening a transaction while nextAddress has
the address buffers mutex currently locked), and instead directly passing it
either:

1. A slice to append future DB address updates, if any
2. The currently open write transaction, if any
3. Neither of the above

In case 1, the database's internal mutexes and never held simultaneously with
the wallet's address mutex locked.

In case 2, the open database transaction locks internal DB mutexes before
locking and unlocking the address mutex.

In case 3, nextAddress is responsible for opening a database transaction and
performing the update.  It now opens this transaction and performs the update
with a deferred function call before locking the address mutex.  Previously,
when the persist callback did not defer the update and did not provide its own
transaction, a new transaction was opening on the spot, with the address
buffers mutex held.  This is incompatible with the lock ordering from case 2,
and would deadlock the wallet.

There is no case 4 where nextAddress is provided both slice to write deferred
updates and an open transaction.  Doing so will return an error with code
errors.Bug.